requirements.txt
appdirs==1.4.3
attrs==19.1.0
backcall==0.1.0
cached-property==1.5.1
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.15.1
lxml==4.4.1
M2Crypto==0.35.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
traitlets==4.3.2
urllib3==1.25.6
wcwidth==0.1.7
zeep==3.4.0
